bench: address CodeRabbit review on PR #26
Fixes 7 of CodeRabbit's 10 inline comments:

- install.sh: flip `&&` to `||` in the apt-get gcc/build-essential path.
  As-was: when c++ was already installed but gcc was missing, pkgs
  started as `(gcc)`, the `&&` arm stripped it (wrong direction), and
  apt-get ran with an empty package list — gcc never installed on
  Debian/Ubuntu hosts that had g++ but not gcc.
- smoke.config.yaml: comment claimed "Smoke runs all four profiles" but
  profiles_enabled lists only two; rewrite to match reality.
- run_mosdepth.smk: comment said `--no-per-base -x` is always used but
  the hybcap branch deliberately omits `-x`. Document why (riker hybcap
  + Picard CHsM both walk CIGARs; mosdepth -x would skip that work and
  make the comparison less fair on capture data).
- stage_inputs.smk: md5 comment claimed "first run populates it" — the
  rule never writes any md5. Rewrite to make the manual-pin workflow
  explicit.
- merge_results.py: `_qualimap_version` final fallback returned
  `lines[0]`, which is qualimap's useless "Java memory size is set to
  ..." banner. Return "unknown" instead.
- select_nygc.py: guard against blank / no-tab lines from `aws s3api`
  output so the script doesn't `ValueError` on a stray header.
- parse_gnu_time.py: parse cpu_percent as float instead of int — GNU
  time outputs whole-number percents in practice but the format isn't
  guaranteed.

Skipped:
- run_riker.smk argv_str:q — would break the rule (params.argv_str is
  the result of shlex.join; adding :q wraps it in single quotes and
  the whole thing becomes a single argv element to `time`, no longer a
  command + args).
- host_info.py re.search guard — overly defensive for a kernel that
  has always emitted a "MemTotal:" line with digits since 2.4.x.
- select_nygc.py auto-pagination — the AWS CLI v2 paginates s3api
  list-objects-v2 by default; the bot's `--no-paginate` suggestion
  would actually CAUSE the truncation it warns about. The script's
  one-time run that locked the sample picks returned all 2504 entries.

+++ benchmark-pipeline/workflow/scripts/parse_gnu_time.py
@@ -43,7 +43,10 @@ def parse(path: str | Path) -> dict:
             elif key == "Maximum resident set size (kbytes)":
                 out["max_rss_kb"] = int(value)
             elif key == "Percent of CPU this job got":
-                out["cpu_percent"] = int(value.rstrip("%"))
+                # `float` rather than `int` since the spec doesn't guarantee
+                # whole-number percents (in practice GNU time emits ints,
+                # but `int("99.5")` would raise on any future change).
+                out["cpu_percent"] = float(value.rstrip("%"))
             elif key == "Exit status":
                 out["exit_status"] = int(value)
     return out
